requirements.txt
Flask==2.3.0
python-dotenv==1.0.0
cerebras-cloud-sdk==1.0.0
pinecone-client==3.0.0
sentence-transformers==2.7.0
numpy<2.0.0
torch>=2.0.0
wikipedia==1.4.0
beautifulsoup4==4.12.0
requests==2.31.0
scikit-learn==1.3.0
